refactor(devices): address review — collapse provider, move hook, rename flag
- Collapse MeetingProviderInner into MeetingProvider (the inner layer only
  existed for the removed remount-key wrapper).
- Move useDeviceController to src/hooks/sdk/ (its own folder no longer holds a
  provider) and delete the empty DeviceControllerProvider folder.
- Rename the internal flag providedDeviceController -> persistDeviceController
  (clearer: the controller persists across leave).
- Document why hostedEventController is captured/restored on leave.

No behavior change; 273 tests pass.
